+++ include/tinytc/tinytc.hpp
@@ -355,13 +355,15 @@ inline data_type make_memref(scalar_type scalar_ty, std::vector<std::int64_t> co
  * @brief Make a group data type
  *
  * @param memref_ty Memref data type
+ * @param offset Offset parameter
  * @param loc Source code location
  *
  * @return Data type
  */
-inline data_type make_group(data_type const &memref_ty, location const &loc = {}) {
+inline data_type make_group(data_type const &memref_ty, std::int64_t offset = 0,
+                            location const &loc = {}) {
     tinytc_data_type_t gt;
-    CHECK_STATUS_LOC(tinytc_group_type_create(&gt, memref_ty.get(), &loc), loc);
+    CHECK_STATUS_LOC(tinytc_group_type_create(&gt, memref_ty.get(), offset, &loc), loc);
     return data_type{gt};
 }