+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.3.0] - 2024-04-29
+
+* Add offset parameter to group type
+* Deploy FindOpenCL and FindLevelZero script for finding tinytc's dependencies via CMake
+
 ## [0.2.0] - 2024-04-26
 
 * Refactored API and added C-API